Add a formula to install python's package installer `pip`
This uses the system's package installer (apt) to install pip, then
we reload salt's modules so the pip module is available to us. We use
`pip` to upgrade itself, and then finally reload the salt modules once
again. This formula will ensure a current version of the pip package
is installed with pip, but it will not update it thereafter.