Add formula to setup and run the hpc-example-war demo (containerized)
This formula will run the hpc-example-war in one of the simplest of
ways, however crude it may be (by salt standards). This formula uses
FPComplete's private docker registry to store a runtime image.

The image was not created with fpbuild, instead using a simple
Dockerfile from the devops-experiments repository. This image was
created on a dev's host, then pushed to the private registry.

The salt master can then instruct the compute nodes to retrieve (pull)
that image and run it as a container. This formula will run the minion
through the steps to pull/run that image/container. There is a separate
formula for the master/slave each. The idea is to use salt's top.sls to
map which hosts get which formula.

As much as is sensible, details have been pulled from salt's pillar,
which allows me to more easily manipulate the details at runtime across
the cluster.