Add a basic stackage.build formula for beta stackage
The 'build' is focused on stackage and stackage-server, and (more
specifically) generating the executables and build artifacts from
these two projects. At present, this is primarily an administrative
tool as it ensures the build artifacts can be copied around in the
cloud (and not have to traverse the net across much slower uplinks).

In the future, we'll connect up Jenkins to one or more of these build
hosts where the builds are actually run.

loose_coupling++

Resolves #4.