Update stackage.server.cabal-loader formula, get functional
This commit gets this formula functional on a host using cron and the
`cabal-loader-stackage` executable. As part of making cabal-loader
functional, the `stackage.server.config` formula has the majority of the
common bits while `stackage.server.init` has stuff specific to the
server, and `stackage.server.cabal-loader` separate as well.

Resolves https://github.com/fpco/devops-experiments/issues/30.