Add formula to install and manage consul, agent/leader roles
This is basic, but rather robust. The goal was to see consul running
on a small cluster, and this formula provides the formula to setup both
leaders and agents. There should be between 3 to 5 leader nodes in a
cluster, per region/datacenter. Any number of agents could then be
connected.

All nodes (leaders/agents) must reference the same `secret_key`.

This formula expects minimal pillar, here is an example:

    ```
    consul:
      leaders:
        - 10.10.2.10
        - 10.10.2.11
        - 10.10.2.12
      leader_count: 3
      secret_key: 'F9fGdI0f+IHrgMP/Eo318g=='
    ```

In this case, the leader nodes must have their `private_ip` set
explicitly, and not just determined at random by the cloud provider.

By default, this formula disables remote execution in consul, as well as
the update check service. All consul nodes have their `node_name` set as
the salt-minion id - grains['id'].

New consul agents will attempt to join the cluster through the leaders,
but no one else. A user and home directory are created for consul.

At present, this formula expects the admin to download the consul .zip
and unpack the executable on the salt-master. In the future we can
improve on the deployment of the executable.

Resolves #10.