Add formula to install the Consul web UI
The formula will ensure the host has the web ui static files unpacked
to the `webui` directory in the consul user home path. For now, the
formula expects the admin to put the webui files on the salt-master.

Also, extend the `consul.agent` formula ever so slightly, running the
`consul agent` in webui mode (pass in the `-ui-dir` option).

Resolves #11.