Add `archive` state module, modified from upstream
TLDR; this verion does not have some of the cruft introduced in the
upstream module, and it actually works.

EG; This is a simplified version of the archive state module (which provides
the `archive.extracted` state). This version is based on an earlier
release of the module (more or less the original pull request), with the
addition of support for the `archive_user` and `keep` options from
upstream.

Moving towards #28 and #16.