Update consul formula to install a specific version from a URL
This update ensures the `consul` executable is installed directly
from a URL. The current setup hardcodes the URL, version, and checksum,
but we can live with that for now.

If the version specified is not installed, the formula will extract the
.zip from the URL to `/usr/local/bin/consul-$version/` and then put a
symlink in place for `/usr/local/bin/consul` to the active version. In
this way, we can gracefully update/rollback, and manage the installed
agent/service with precision.

Resolves #41.