Add formula to install and configure scalyr-agent-2
This is as simple as it can get, while being totally focused on this
service as it is used by stackage.org right now (in trial).

This formula is not generic enough, but it should be committed before
it is lost. The config JSON is embedded in `scalyr/init.sls`.

The installation process is based on the details provided in their docs
https://www.scalyr.com/help/install-agent

Resolves #56.