Add formula to install and configure nginx
This formula was inspired by the `ng` formula seen in the SaltStack
Formulas nginx repo seen here:
https://github.com/saltstack-formulas/nginx-formula

This incarnation is far simpler and more readable, forgoing support
for what we don't need (most everything outside Ubuntu install and
configuration). Set the stage for vhost config, but don't iterate
over the vhost setup. Put as much as possible in map.jinja, and in
pillar. Merge it all together to keep it usable.

Resolves #57.