Add formula to drop default user
While yes, we can do this with pillar and the existing `users` formula,
it is nice to have a formula that specifically does this (drop the
default user from an image/host once provisioned).

Ensure we also rm the default user's 90-cloud-init-users sudoers config,
and purge (rm) the user's home directory.

Resolves #53.