Bugfix in storage.lvm formula
The latest release of Saltstack (2015.5.2 as of now) parses a hanging
comma in the devices parameter as an empty string, and we're left with
an empty element in the array passed to the lvm state called (which
errors out). This patch ensures the comma is only added if there is
more than one device to be printed in the `devices` list.