Add formula to integrate consul with users formula, via consul-template
This provides consul-template with the config to watch the `users` key
in consul for changes. This key ought to be populated with the pillar
key for the users salt formula. Raw, YAML-formatted string, and
consul-template will dump it to /srv/pillar/users.sls. If the host's
pillar top.sls includes the users definition, admins can manage the
user formula via the webui.

Resolves DEVOPS-272.