create a jinja macro to generate states for monitoring consul
This macro is basically a template made from the generic version of
the pattern developed to integrate consul and salt formula (without
having salt pull pillar directly from consul). The formula are applied
in the masterless mode when the monitored keys in consul change and the
template is rendered. The macro here makes a DRY version available for
the generic use case, which we exemplify by using the macro in the users
consul-template formula.

These monitors are the consul/salt formula integration.