Update docker.compose formula with a lookup map for checksums
define a map, version to sha512 checksum hash for the docker-compose
executable downloaded from github.

this simplifies use of the formula - the opsdev can now specify a
version, and the formula will lookup the checksum for that version
of the binary, as downloaded from github. A custom checksum can be
specified if the user wishes to do so.