set nomad server's consul agent check to use service status
The idea is that a) when the cluster does not have consensus on a
leader, the HTTP check fails, and b) when the consul agent check
fails, the node is taken out of service in consul.. and so nomad
would need manual intervention to form consensus on a leader. Now, if
the nomad server is running, consul puts it in service, which puts nomad
on the map again (in DNS).. the server may not be 100% functional, but
the nomad servers will find each other and form consensus on their own
this way.. and we'll setup the nomad agents to retry until they get
connected to the working nomad server cluster.