python/pip: pin version of pip installed
Update the formula so we can pin the version, to be more
specific about what we install.

Use pillar to allow for overriding that version specified.
Use the `pip_version` key, and provide a default in case
there is no value for that pillar key.

Set the default to "less than 10.x", as the 10.x series
breaks saltstack. Resolves #120.

See also saltstack/salt#46163 for more info on all of this.