reclass: install fork from git, this is "1.4.2"
This is for #137. The fork ports reclass to latest python, addresses
some issues, and is discussed in general here:

https://github.com/saltstack/salt/issues/47126

Commit history for the fork installed by default can be seen in:

https://github.com/max-arnold/reclass/commits/maintenance

And the git repo/fork used can be overridden via pillar:

`reclass:git_url` is the key. To specify a branch on a repo, use the
form: git+https://github.com/user/repo.git@branch