vagrant/tests: use `ops http poll` to address "the waiting problem"
that problem is: how long will it take to download and run the task?

the time to do so varies from host to host and from time to time, and
when we don't want to wait longer than we need to.. and we don't want to
have a false positive saying the service is unhealthy when it just
needed another few seconds to come online.. thus, we use an HTTP poll to
confirm when we should proceed. We don't want that to poll forever, so
we set 60 seconds as an upper limit. Update the comments while we're
here.. we have the echo's to cover.