vagrant: add support for building a VM to package as a .box for import
Vagrant runs VMs with "boxes", and `vagrant package` will create a
`.box` file (export) from a running vagrant VM. This `Vagrantfile` will
create a VM that has applied the "install" portion of the saltstack
formula in this repo.

This generates a "foundation" we can use to more quickly and easily run
new VMs with specific roles (apply the "config and run" portion of the
formula).